Extract native CLI setup command execution
Move install and uninstall process execution into a focused runner so the
installer remains responsible for orchestration and current-process state.
Keep command behavior and timeout handling unchanged while routing existing
tests directly through the extracted owner.
